Update cookies checks in example
I've found out that when there are no other cookies, checking just for a truthy value of `cookie` doesn't seem to work, but checking it is an actual string as in the example for older Next versions seems to work fine. I was wondering if this change was intentional, or just a mistake. 

Please, feel free to close this PR if the change was intentional and this is not the problem. 
Thanks!