removes unnecessary ARIA
The button to open the search text field had a `role=search` on it, as well as an `aria-label`.

[`role=search` is a type of landmark](https://www.w3.org/TR/wai-aria-1.2/#search), and not a role that indicates it is actionable.    Removing this role allows the control to be correctly exposed as a button element.

The `aria-label="Search the docs"` on the button suppresses the announcement of its child content, making it so that the key commands to open the search modal will not be announced.  

As "Search the docs" is already child content of the button (line 67), removing the `aria-label` allows the name of this button to still be announced correctly, and the additional "press ... to search" content will now be announced.

If you would like the search landmark to still be used / exposed, it must be an element that wraps this button, and not used on the button itself.